+++ .swiftpm/xcode/xcuserdata/filippo.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -7,15 +7,12 @@
 		<key>Kovalee-iOS-SDK-Package.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-<<<<<<< HEAD
 			<integer>6</integer>
 		</dict>
 		<key>KovaleeAds.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-=======
->>>>>>> 205cf3c (feat: moved Adjust to new target KovaleeAttribution)
-			<integer>3</integer>
+			<integer>4</integer>
 		</dict>
 		<key>KovaleeAttribution.xcscheme_^#shared#^_</key>
 		<dict>
@@ -25,7 +22,7 @@
 		<key>KovaleePurchases.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>4</integer>
+			<integer>2</integer>
 		</dict>
 		<key>KovaleeRemoteConfig.xcscheme_^#shared#^_</key>
 		<dict>
@@ -40,62 +37,42 @@
 		<key>KovaleeSDK.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>2</integer>
+			<integer>3</integer>
 		</dict>
 		<key>Promises (Playground) 1.xcscheme</key>
 		<dict>
 			<key>isShown</key>
 			<false/>
 			<key>orderHint</key>
-<<<<<<< HEAD
 			<integer>9</integer>
-=======
-			<integer>6</integer>
->>>>>>> 205cf3c (feat: moved Adjust to new target KovaleeAttribution)
 		</dict>
 		<key>Promises (Playground) 2.xcscheme</key>
 		<dict>
 			<key>isShown</key>
 			<false/>
 			<key>orderHint</key>
-<<<<<<< HEAD
 			<integer>10</integer>
-=======
-			<integer>7</integer>
->>>>>>> 205cf3c (feat: moved Adjust to new target KovaleeAttribution)
 		</dict>
 		<key>Promises (Playground) 3.xcscheme</key>
 		<dict>
 			<key>isShown</key>
 			<false/>
 			<key>orderHint</key>
-<<<<<<< HEAD
 			<integer>11</integer>
-=======
-			<integer>8</integer>
->>>>>>> 205cf3c (feat: moved Adjust to new target KovaleeAttribution)
 		</dict>
 		<key>Promises (Playground) 4.xcscheme</key>
 		<dict>
 			<key>isShown</key>
 			<false/>
 			<key>orderHint</key>
-<<<<<<< HEAD
 			<integer>12</integer>
-=======
-			<integer>9</integer>
->>>>>>> 205cf3c (feat: moved Adjust to new target KovaleeAttribution)
 		</dict>
 		<key>Promises (Playground) 5.xcscheme</key>
 		<dict>
 			<key>isShown</key>
 			<false/>
 			<key>orderHint</key>
-<<<<<<< HEAD
 			<integer>13</integer>
-=======
-			<integer>10</integer>
->>>>>>> 205cf3c (feat: moved Adjust to new target KovaleeAttribution)
 		</dict>
 		<key>Promises (Playground) 6.xcscheme</key>
 		<dict>
@@ -123,11 +100,7 @@
 			<key>isShown</key>
 			<false/>
 			<key>orderHint</key>
-<<<<<<< HEAD
 			<integer>8</integer>
-=======
-			<integer>5</integer>
->>>>>>> 205cf3c (feat: moved Adjust to new target KovaleeAttribution)
 		</dict>
 	</dict>
 	<key>SuppressBuildableAutocreation</key>
@@ -137,20 +110,16 @@
 			<key>primary</key>
 			<true/>
 		</dict>
-<<<<<<< HEAD
 		<key>KovaleeAds</key>
 		<dict>
 			<key>primary</key>
 			<true/>
 		</dict>
-=======
->>>>>>> 205cf3c (feat: moved Adjust to new target KovaleeAttribution)
 		<key>KovaleeAttribution</key>
 		<dict>
 			<key>primary</key>
 			<true/>
 		</dict>
-<<<<<<< HEAD
 		<key>KovaleePurchases</key>
 		<dict>
 			<key>primary</key>
@@ -161,8 +130,6 @@
 			<key>primary</key>
 			<true/>
 		</dict>
-=======
->>>>>>> 205cf3c (feat: moved Adjust to new target KovaleeAttribution)
 		<key>KovaleeSDK</key>
 		<dict>
 			<key>primary</key>
